track delete test

+++ src/main/java/edu/montana/csci/csci440/model/Track.java
@@ -47,7 +47,7 @@ public class Track extends Model {
         genreId = results.getLong("GenreId");
     }
 
-    public static Track find(int i) {
+    public static Track find(long i) {
         try (Connection conn = DB.connect();
              PreparedStatement stmt = conn.prepareStatement("SELECT * FROM tracks WHERE TrackId=?")) {
             stmt.setLong(1, i);
